Move admin functions to `admin.js`
